requirements.txt
numpy>=1.21.0
pandas>=1.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.0.0
joblib>=1.1.0
